fix(lsp): ignore diagnostics from a LSP we're detached from
This race condition can happen when a LSP is emitting metrics
while we're detaching the client.

How to reproduce:

- Use a neovim with https://github.com/neovim/neovim/pull/33707, and a
  Nix LSP such as `nixd`.
- Open `myscript`:
  ```python
  #!/usr/bin/env nix-shell
  #!nix-shell -i python3 -p python3

  # vi: ft=python

  print("Hello, world!")
  ```
- You should see a clean buffer (no diagnostics)
- Now, reopen the file with `:e`. You should see a cached diagnostic from
  `nixd` complaining about the `print` statement.

I intend to submit this patch to Neovim once the above PR has landed. I don't
want to try to explain it before then, as it's quite a weird situation, and I
need someone more familiar with Neovim's LSP/diagnostics architecture to tell
me how to test this and if there's a better way to implement the fix.

Also see this comment in `lsp.buf_detach_client`:
https://github.com/neovim/neovim/blob/v0.11.1/runtime/lua/vim/lsp.lua#L971-L972